Changed controls of spaceship . New controls : W (Move forward) , A(Turn counter clockwise), D(Turn clockwise) and S(Move back).

+++ src/Drawable.h
@@ -12,11 +12,11 @@ class Drawable : public GameObject
 {
 	private:
 		ALLEGRO_BITMAP* img;
-		int coordinates[8];
-		unsigned int length;
-		unsigned int height;
-		int x;
-		int y;
+		float coordinates[8];
+		float length;
+		float height;
+		float x;
+		float y;
 		float angle;
 		string path;
 	public:
@@ -24,19 +24,19 @@ class Drawable : public GameObject
 		Drawable();
 
 		//constructor . This constructor initialize all the class variables using the input values . path is the path to the image of this drawable object
-		Drawable(unsigned int length, unsigned int height, int x, int y, string path, float angle);
+		Drawable(float length, float height, float x, float y, string path, float angle);
 
 		//destructor
 		~Drawable();
 
 		//gets the new x and y and updates the coordinates.
-		void UpdateCoordinates(int x, int y);
+		void UpdateCoordinates(float x, float y);
 
 		//changes the value of class variable x
-		void SetX(int x);
+		void SetX(float x);
 
 		//changes the value of class variable y
-		void SetY(int y);
+		void SetY(float y);
 
 		//changes the value of class variable angle
 		void SetAngle(float angle);
@@ -45,10 +45,10 @@ class Drawable : public GameObject
 		float GetAngle();
 
 		//returns the class variable x
-		int GetX();
+		float GetX();
 
 		//returns the class variable y
-		int GetY();
+		float GetY();
 		
 		//draws the object with al_draw_bitmap function
 		void Draw();
