+++ src/Moveable.h
@@ -24,7 +24,7 @@ class Moveable : public Drawable
 		Moveable();
 
 		//Constructor . This constructor initialize all the class (and its father's) variables using the input values . path is the path to the image of this drawable object
-		Moveable(int speed, unsigned int length, unsigned int height, int x, int y, string path, float angle);
+		Moveable(int speed, float length, float height, float x, float y, string path, float angle);
 
 		//Destructor default
 		~Moveable();
