+++ src/Spaceship.cpp
@@ -5,7 +5,7 @@ Spaceship::Spaceship() : Moveable()
 	this->lives = 3;
 }
 
-Spaceship::Spaceship(unsigned int lives, int speed, unsigned int length, unsigned int height, int x, int y, string path, float angle) : Moveable(speed, length, height, x, y, path, angle)
+Spaceship::Spaceship(unsigned int lives, int speed, float length, float height, float x, float y, string path, float angle) : Moveable(speed, length, height, x, y, path, angle)
 {
 	this->lives = lives;
 }
