+++ src/Spaceship.h
@@ -12,7 +12,7 @@ class Spaceship : public Moveable
 		Spaceship();
 
 		//Constructor . This constructor initialize all the class (and its father's) variables using the input values . path is the path to the image of this drawable object
-		Spaceship(unsigned int lives, int speed, unsigned int length, unsigned int height, int x, int y, string path, float angle);
+		Spaceship(unsigned int lives, int speed, float length, float height, float x, float y, string path, float angle);
 
 		//Destructor
 		~Spaceship();
